# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Stash-box metadata database",
        "title": "Stash-box REST API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Stash App",
            "url": "https://github.com/stashapp/stash-box"
        },
        "license": {
            "name": "AGPL-3.0",
            "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/scenes/fingerprints/batch": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth SessionAuth": []
                    }
                ],
                "tags": [
                    "scenes"
                ],
                "summary": "Match multiple fingerprints in batch (Plex critical)",
                "parameters": [
                    {
                        "description": "Array of fingerprint arrays (one array per scene)",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_rest_handlers_scene.FingerprintBatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "array",
                                                "items": {
                                                    "$ref": "#/definitions/github_com_stashapp_stash-box_internal_models.Scene"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "413": {
                        "description": "Request Entity Too Large",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/scenes/fingerprints/match": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth SessionAuth": []
                    }
                ],
                "tags": [
                    "scenes"
                ],
                "summary": "Match a single fingerprint",
                "parameters": [
                    {
                        "description": "Fingerprint data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_rest_handlers_scene.FingerprintMatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_stashapp_stash-box_internal_models.Scene"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/scenes/{id}/fingerprints": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth SessionAuth": []
                    }
                ],
                "tags": [
                    "scenes"
                ],
                "summary": "Submit a fingerprint for a scene",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Scene ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fingerprint to submit",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_rest_handlers_scene.SubmitFingerprintRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "boolean"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_gofrs_uuid.NullUUID": {
            "type": "object",
            "properties": {
                "uuid": {
                    "type": "string"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
        "github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_stashapp_stash-box_internal_api_rest_handlers.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {}
            }
        },
...
        "github_com_stashapp_stash-box_internal_models.Scene": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "deleted": {
                    "type": "boolean"
                },
                "details": {
                    "type": "string"
                },
                "director": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "production_date": {
                    "type": "string"
                },
                "studio_id": {
                    "$ref": "#/definitions/github_com_gofrs_uuid.NullUUID"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "internal_api_rest_handlers_scene.FingerprintBatchRequest": {
            "type": "object",
            "required": [
                "fingerprints"
            ],
            "properties": {
                "fingerprints": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/internal_api_rest_handlers_scene.FingerprintQuery"
                        }
                    }
                }
            }
        },
...
        "internal_api_rest_handlers_scene.FingerprintMatchRequest": {
            "type": "object",
            "required": [
                "algorithm",
                "hash"
            ],
            "properties": {
                "algorithm": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                }
            }
        },
        "internal_api_rest_handlers_scene.FingerprintQuery": {
            "type": "object",
            "required": [
                "algorithm",
                "hash"
            ],
            "properties": {
                "algorithm": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                }
            }
        },
        "internal_api_rest_handlers_scene.SubmitFingerprintRequest": {
            "type": "object",
            "required": [
                "algorithm",
                "hash"
            ],
            "properties": {
                "algorithm": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "hash": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Key authentication",
            "type": "apiKey",
            "name": "ApiKey",
            "in": "header"
        },
        "SessionAuth": {
            "description": "Session cookie authentication",
            "type": "apiKey",
            "name": "stashbox",
            "in": "cookie"
        }
    }
}